BreakingShadowDjango/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==11.0.0
python-dotenv==1.0.1
sqlparse==0.5.1
